rfc6979

Function bits2int

pub(crate) fn bits2int<U>(b: &[u8], q: &U) -> U
where
    U: Unsigned + Encoding,

bits2int transform as defined in RFC6979 2.3.2: “Bit String to Integer”

From the RFC:

The bits2int transform takes as input a sequence of blen bits and outputs a non-negative integer that is less than 2^qlen. It consists of the following steps:

The sequence is first truncated or expanded to length qlen:

if qlen < blen, then the qlen leftmost bits are kept, and subsequent bits are discarded;

otherwise, qlen-blen bits (of value zero) are added to the left of the sequence (i.e., before the input bits in the sequence order).

The resulting sequence is then converted to an integer value using the big-endian convention: if input bits are called b_0 (leftmost) to b_(qlen-1) (rightmost), then the resulting value is:

b_02^(qlen-1) + b_12^(qlen-2) + … + b_(qlen-1)*2^0

The bits2int transform can also be described in the following way: the input bit sequence (of length blen) is transformed into an integer using the big-endian convention. Then, if blen is greater than qlen, the resulting integer is divided by two to the power blen-qlen (Euclidian division: the remainder is discarded); in many software implementations of arithmetics on big integers, that division is equivalent to a “right shift” by blen-qlen bits.