rfc6979

Function bits2octets

pub(crate) fn bits2octets<'a, U>(b: &[u8], q: &U, out: &'a mut [u8]) -> &'a [u8] 
where
    U: Unsigned + Encoding,

bits2octets transform as defined in [RFC6979 2.3.4]: “Bit String to Octet String”

From the RFC:

The bits2octets transform takes as input a sequence of blen bits and outputs a sequence of rlen bits. It consists of the following steps:

The input sequence b is converted into an integer value z1 through the bits2int transform:

z1 = bits2int(b)

z1 is reduced modulo q, yielding z2 (an integer between 0 and q-1, inclusive):

z2 = z1 mod q

Note that since z1 is less than 2^qlen, that modular reduction can be implemented with a simple conditional subtraction: z2 = z1-q if that value is non-negative; otherwise, z2 = z1.

z2 is transformed into a sequence of octets (a sequence of rlen bits) by applying int2octets.