rfc6979

Function int2octets

pub(crate) fn int2octets<'a, U: Unsigned>(
    x: &U,
    q: &U,
    out: &'a mut [u8],
) -> &'a [u8] 

int2octets transform as defined in RFC6979 2.3.3: “Integer to Octet String”

From the RFC:

An integer value x less than q (and, in particular, a value that has been taken modulo q) can be converted into a sequence of rlen bits, where rlen = 8*ceil(qlen/8). This is the sequence of bits obtained by big-endian encoding. In other words, the sequence bits x_i (for i ranging from 0 to rlen-1) are such that:

x = x_02^(rlen-1) + x_12^(rlen-2) + … + x_(rlen-1)

We call this transform int2octets. Since rlen is a multiple of 8 (the smallest multiple of 8 that is not smaller than qlen), then the resulting sequence of bits is also a sequence of octets, hence the name.